2008-11-18  vadim  <vadim@vk.crocodile.org>
* prefsdialog_q.ui: better layout of the first page of Preferences
dialog to make sure long path to the working directory fits in the
input widget.

* SSHPIX.cpp (SSHPIX::stateMachine): bugfix: installer for Cisco
routers and PIX could not find generated file because variable
conffile is now always a full absolute path. This bug was
introduced earlier during installer rewrite for v3.0.2. Tested
installer for router and PIX using default generated file name, as
well as custom generated file name, defined both as absolute and
as relative path. Tested batch install of combination of a router
and a pix in one batch (the same user account, then same enable
password on both)