2008-11-22  vadim  <vadim@vk.crocodile.org>
* StartTipDialog.cpp (StartTipDialog::StartTipDialog): Added
"start tip" dialog that shows brief information on the online
resources available to the user (web site URL, links to the
Firewall Builder FAQ, HOWTOs, Cook Book). Linked pages open in the
standard browser.

* FWWindow.cpp: added menu item Help/Firewall Builder Help that
opens a page with information about online resources for Firewall
Builder (the same page that is shown in the "start tip" dialog).