2008-12-25  vadim  <vadim@vk.crocodile.org>
* Compiler.cpp (Compiler::_expandAddressRanges): using
FWObjectDatabase::createClass methods to create rules and other
objects in compilers wherever the type is known at the (code)
compile time. This makes code cleaner and speeds it up a little
because of eliminated cast() and string comparison.

2008-12-24  vadim  <vadim@vk.crocodile.org>

* FWObjectDatabase_create_object.cpp (init_create_methods_table):
using table of pointers to methods that create objects of
different types to dispatch call to createFWObjectClass methods
and avoid long chains of "if" operators that compare type names.

* FWObjectDatabase_create_object.cpp (CALL_CREATE_OBJ_METHOD):
added bunch of FWObjectDatabase::createClass methods for all known
object types to create objects of the corresponding class and
return a pointer to it. This provides a way to create an object
without excessive use of dynamic_cast() and many if operators to
compare type name.