C++ exception specifications are deprecated in C++11:
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2010/n3051.html

C++ exceptions are dynamic and evaluated at runtime. This greatly
limits the value of this feature, and introduces a performance
runtime overhead to perform the evaluations.

Recent versions of GCC warn about use of this deprecated feature.
Instead of disabling the warning, this change eliminates the use
of this deprecated feature.

There were a few places where "throw()" was retained as the
classes extend the standard library and must either specify
"throw()" or "noexcept" in order to meet the interface
contract. As "noexcept" is not backwards compatible,
"throw()" has been retained.