Add GCC7 version directives to avoid __attribute__ errors on fallthrough
Building with ealier GCC (e.g. 5.4.0 on Ubuntu 16.04) causes compiler
errors during the build due to the fallthrough warning capabilities
added in GCC7 described in:

https://developers.redhat.com/blog/2017/03/10/wimplicit-fallthrough-in-gcc-7/

The patch follows the approach from https://stackoverflow.com/a/53030657/8928529
to add directives to wrap the __attribute__ call and avoid failure on earlier
GCC versions.

An alternate solution could be to use a fall through comment for greater
portability. https://stackoverflow.com/a/45137452/8928529