Set a white text color if the background is dark
This checks that the "grayness" of the background is less than 100
and switches to use white for the text color if so.

Testing against a dark theme here suggests that this helps a lot. The
"grayness" for a dark theme is typically less than 60, but contrast issues
should start to disappear around 100.