For dark theme also rewrite Standard library color to dark blue
This changes the background colour for the Standard library to dark blue,
even when loading from file.  Potentially this could check for the
standard #d4f8ff value before doing this in case someone had manually
customised their `objects_init.xml` template but I think that would be
better to wait for a proper theming solution.