Avoid re-creating of "fwb_session_with_keepalive".
On Windows, that session is created by the installer already with the correct values, so can easily be assumed being available. OTOH, some people don't use PLINK/PSCP with FWB at all and might delete the created session by the installer to not bloat their otherwise used PuTTY-sessions. If done so, FWB re-creating that session on every start is a bit annoying for those people.

This fixes https://github.com/fwbuilder/fwbuilder/issues/120.