Working rpmbuild for Fedora
Changes needed for rpmbuild on Fedora to work with version 6.0.0.  This
has only been tested on Fedora 34, but it should be close for Fedora 35+
and (possibly) current versions of CentOS.

Use cmake to generate the RPM spec file and a tarball creation script
with the correct version number taken from the VERSION cmake script.
From there, the user can follow the instructions to create a source RPM
package and run rpmbuild to create the binary packages.